requirements.txt
opencv-python
numpy
pyzbar
fastapi
uvicorn
python-multipart
pillow